BPAAS-974: Fix variable names
