+++ api/handler/default.go
@@ -565,27 +565,27 @@ func (m ArcDefaultHandler) getTransaction(ctx context.Context, inputTxID string)
 func (m ArcDefaultHandler) getPolicyFromNode() (*bitcoin.Settings, error) {
 	peerRpcPassword := viper.GetString("peerRpc.password")
 	if peerRpcPassword == "" {
-		return nil, errors.Errorf("setting peerRpcPassword not found")
+		return nil, errors.Errorf("setting peerRpc.password not found")
 	}
 
 	peerRpcUser := viper.GetString("peerRpc.user")
 	if peerRpcUser == "" {
-		return nil, errors.Errorf("setting peerRpcUser not found")
+		return nil, errors.Errorf("setting peerRpc.user not found")
 	}
 
 	peerRpcHost := viper.GetString("peerRpc.host")
 	if peerRpcHost == "" {
-		return nil, errors.Errorf("setting peerRpcHost not found")
+		return nil, errors.Errorf("setting peerRpc.host not found")
 	}
 
 	peerRpcPort := viper.GetInt("peerRpc.port")
 	if peerRpcPort == 0 {
-		return nil, errors.Errorf("setting peerRpcPort not found")
+		return nil, errors.Errorf("setting peerRpc.port not found")
 	}
 
 	rpcURL, err := url.Parse(fmt.Sprintf("rpc://%s:%s@%s:%d", peerRpcUser, peerRpcPassword, peerRpcHost, peerRpcPort))
 	if err != nil {
-		return nil, errors.Errorf("failed to parse peerRpcHost and peerRpcPort: %v", err)
+		return nil, errors.Errorf("failed to rpc URL: %v", err)
 	}
 
 	// connect to bitcoin node and get the settings
