+++ blocktx/BlockNotifier.go
@@ -1,9 +1,6 @@
 package blocktx
 
 import (
-	"fmt"
-	"net/url"
-
 	"github.com/bitcoin-sv/arc/blocktx/blocktx_api"
 	"github.com/bitcoin-sv/arc/blocktx/store"
 	"github.com/libsv/go-p2p"
@@ -29,52 +26,6 @@ type BlockNotifier struct {
 	deadSubscriptions chan subscriber
 	blockCh           chan *blocktx_api.Block
 	quitCh            chan bool
-}
-
-type Peer struct {
-	Host string
-	Port PeerPort `mapstructure:"port"`
-}
-
-type PeerPort struct {
-	P2P int `mapstructure:"p2p"`
-	ZMQ int `mapstructure:"zmq"`
-}
-
-func (p Peer) GetZMQUrl() (*url.URL, error) {
-	if p.Port.ZMQ == 0 {
-		return nil, fmt.Errorf("port_zmq not set for peer %s", p.Host)
-	}
-
-	if p.Host == "" {
-		return nil, fmt.Errorf("host not set for peer %s", p.Host)
-	}
-
-	zmqURLString := fmt.Sprintf("zmq://%s:%d", p.Host, p.Port.ZMQ)
-
-	return url.Parse(zmqURLString)
-}
-
-func (p Peer) GetP2PUrl() (string, error) {
-	if p.Port.P2P == 0 {
-		return "", fmt.Errorf("port_p2p not set for peer %s", p.Host)
-	}
-
-	if p.Host == "" {
-		return "", fmt.Errorf("host not set for peer %s", p.Host)
-	}
-
-	return fmt.Sprintf("%s:%d", p.Host, p.Port.P2P), nil
-}
-
-func GetPeerSettings() ([]Peer, error) {
-	var peers []Peer
-	err := viper.UnmarshalKey("peers", &peers)
-	if err != nil {
-		return []Peer{}, err
-	}
-
-	return peers, nil
 }
 
 func NewBlockNotifier(storeI store.Interface, l utils.Logger) *BlockNotifier {
